requirements.txt
streamlit
google-genai
firebase-admin
Pillow